move ensure module and ArithmeticError to sp-arithmetic
